Fix non-dynamic revalidations. Change post page to ISR

+++ packages/server/src/graphql/nexus/mutations/deleteExperience.ts
@@ -25,12 +25,14 @@ export const deleteExperience = mutationField("deleteExperience", {
 
 		return false;
 	},
-	resolve: async (parent, args, { prisma, res }) => {
+	resolve: async (parent, args, { prisma, res, user }) => {
+		if (!user) throw new Error();
+
 		const record = await prisma.experience.delete({
 			where: PrismaUtils.nonNull(args.where)
 		});
 
-		await res.unstable_revalidate("/leedavidcs").catch(() => null);
+		await res.unstable_revalidate(`/${user.name}`).catch(() => null);
 
 		return { record };
 	}
